# server/router/settings-router.go
package router

import (
	"encoding/json"
	"errors"
	"log"
	"net/http"
	"strconv"

	"github.com/gorilla/mux"

	. "github.com/seatsurfing/seatsurfing/server/api"
	. "github.com/seatsurfing/seatsurfing/server/config"
	"github.com/seatsurfing/seatsurfing/server/plugin"
	. "github.com/seatsurfing/seatsurfing/server/repository"
	. "github.com/seatsurfing/seatsurfing/server/util"
)

type SettingsRouter struct {
}

type SetSettingsRequest struct {
	Value string `json:"value" validate:"max=256"`
}

type GetSettingsResponse struct {
	Name  string `json:"name"`
	Value string `json:"value"`
}

type OfficeSettingsResponse struct {
	WorkStartTime string `json:"workStartTime"`
	WorkEndTime   string `json:"workEndTime"`
}

type SettingsRouterAdminMenuItem struct {
	ID         string `json:"id"`
	Title      string `json:"title"`
	Source     string `json:"src"`
	Visibility string `json:"visibility"`
	Icon       string `json:"icon"`
}

type SettingsRouterWelcomeScreen struct {
	Source string `json:"src"`
}

var (
	ErrAlreadyExists               = errors.New("resource already exists")
	SysSettingOrgSignupDelete      = "_sys_org_signup_delete"
	SysSettingVersion              = "_sys_version"
	SysSettingAdminMenuItems       = "_sys_admin_menu_items"
	SysSettingAdminWelcomeScreens  = "_sys_admin_welcome_screens"
	SysSettingOrgPrimaryDomain     = "_sys_org_primary_domain"
	SysSettingDisablePasswordLogin = "_sys_disable_password_login"
)

func (router *SettingsRouter) SetupRoutes(s *mux.Router) {
	s.HandleFunc("/timezones", router.getTimezones).Methods("GET")
	s.HandleFunc("/office-hours", router.getOfficeHours).Methods("GET")
	s.HandleFunc("/office-hours", router.setOfficeHours).Methods("PUT")
	s.HandleFunc("/{name}", router.getSetting).Methods("GET")
	s.HandleFunc("/{name}", router.setSetting).Methods("PUT")
	s.HandleFunc("/", router.getAll).Methods("GET")
	s.HandleFunc("/", router.setAll).Methods("PUT")
}

func (router *SettingsRouter) getTimezones(w http.ResponseWriter, r *http.Request) {
	SendJSON(w, TimeZones)
}

func (router *SettingsRouter) getOfficeHours(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !CanAdminOrg(user, user.OrganizationID) {
		SendForbidden(w)
		return
	}
	officeSettings, err := GetOfficeSettingsRepository().Get()
	if err != nil {
		if errors.Is(err, ErrOfficeSettingsMissing) {
			SendNotFound(w)
			return
		}
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	SendJSON(w, router.copyOfficeSettingsToRestModel(officeSettings))
}

func (router *SettingsRouter) setOfficeHours(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !CanAdminOrg(user, user.OrganizationID) {
		SendForbidden(w)
		return
	}
	var value OfficeSettingsResponse
	if UnmarshalBody(r, &value) != nil {
		SendBadRequest(w)
		return
	}
	officeSettings := &OfficeSettings{
		WorkStartTime: value.WorkStartTime,
		WorkEndTime:   value.WorkEndTime,
	}
	if err := GetOfficeSettingsRepository().Upsert(officeSettings); err != nil {
		SendBadRequest(w)
		return
	}
	SendUpdated(w)
}

func (router *SettingsRouter) getSetting(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	vars := mux.Vars(r)
	orgAdmin := CanAdminOrg(user, user.OrganizationID)
	if !((orgAdmin && router.isValidSettingNameReadAdmin(vars["name"])) || (router.isValidSettingNameReadPublic(vars["name"]))) {
		SendForbidden(w)
		return
	}
	if (vars["name"] == SysSettingOrgSignupDelete) && orgAdmin {
		SendJSON(w, router.getSysSettingOrgSignupDelete())
		return
	}
	if vars["name"] == SysSettingVersion {
		SendJSON(w, router.getSysSettingVersion())
		return
	}
	if vars["name"] == SysSettingAdminMenuItems {
		SendJSON(w, router.getAdminMenuItems())
		return
	}
	if vars["name"] == SysSettingAdminWelcomeScreens {
		list, _ := GetSettingsRepository().GetAll(user.OrganizationID)
		SendJSON(w, router.getAdminWelcomeScreens(list))
		return
	}
	if vars["name"] == SysSettingOrgPrimaryDomain {
		sysSettingOrgPrimaryDomain := router.getSysSettingOrgPrimaryDomain(user.OrganizationID)
		SendJSON(w, sysSettingOrgPrimaryDomain.Value)
		return
	}
	if vars["name"] == SysSettingDisablePasswordLogin {
		sysSettingDisablePasswordLogin := router.getSysSettingDisablePasswordLogin()
		SendJSON(w, sysSettingDisablePasswordLogin.Value)
		return
	}
	value, err := GetSettingsRepository().Get(user.OrganizationID, vars["name"])
	if err != nil {
		log.Println(err)
		SendNotFound(w)
		return
	}
	SendJSON(w, value)
}

func (router *SettingsRouter) setSetting(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !CanAdminOrg(user, user.OrganizationID) {
		SendForbidden(w)
		return
	}
	var value SetSettingsRequest
	if UnmarshalValidateBody(r, &value) != nil {
		SendBadRequest(w)
		return
	}
	vars := mux.Vars(r)
	if !router.isValidSettingNameWrite(vars["name"]) {
		SendNotFound(w)
		return
	}
	if !router.isValidSettingType(vars["name"], value.Value) {
		SendBadRequest(w)
		return
	}
	if !router.isValidSettingValue(vars["name"], value.Value) {
		SendBadRequest(w)
		return
	}
	err := router.doSetOne(user.OrganizationID, vars["name"], value.Value)
	if err != nil {
		log.Println(err)
		if errors.Is(err, ErrAlreadyExists) {
			SendAlreadyExists(w)
		} else {
			SendInternalServerError(w)
		}
		return
	}
	SendUpdated(w)
}

func (router *SettingsRouter) getAll(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !CanAccessOrg(user, user.OrganizationID) {
		SendForbidden(w)
		return
	}
	orgAdmin := CanAdminOrg(user, user.OrganizationID)
	list, err := GetSettingsRepository().GetAll(user.OrganizationID)
	if err != nil {
		log.Println(err)
		SendInternalServerError(w)
		return
	}
	res := []*GetSettingsResponse{}
	for _, e := range list {
		if (orgAdmin && router.isValidSettingNameReadAdmin(e.Name)) || (router.isValidSettingNameReadPublic(e.Name)) {
			m := router.copyToRestModel(e)
			res = append(res, m)
		}
	}
	if orgAdmin {
		res = append(res, router.getSysSettingOrgSignupDelete())
		res = append(res, router.getAdminWelcomeScreens(list))
	}
	if CanSpaceAdminOrg(user, user.OrganizationID) {
		res = append(res, router.getAdminMenuItems())
	}
	res = append(res, router.getSysSettingVersion())
	res = append(res, router.getSysSettingOrgPrimaryDomain(user.OrganizationID))
	res = append(res, router.getSysSettingDisablePasswordLogin())
	for _, plg := range plugin.GetPlugins() {
		plgSettings := (*plg).GetPublicSettings(user.OrganizationID)
		for _, setting := range plgSettings {
			res = append(res, &GetSettingsResponse{
				Name:  setting.Name,
				Value: setting.Value,
			})
		}
	}
	SendJSON(w, res)
}

func (router *SettingsRouter) setAll(w http.ResponseWriter, r *http.Request) {
	user := GetRequestUser(r)
	if !CanAdminOrg(user, user.OrganizationID) {
		SendForbidden(w)
		return
	}
	var list []GetSettingsResponse
	if err := UnmarshalBody(r, &list); err != nil {
		log.Println(err)
		SendBadRequest(w)
		return
	}
	for _, e := range list {
		if !router.isValidSettingNameWrite(e.Name) {
			SendNotFound(w)
			return
		}
		if !router.isValidSettingType(e.Name, e.Value) {
			SendBadRequest(w)
			return
		}
		if !router.isValidSettingValue(e.Name, e.Value) {
			SendBadRequest(w)
			return
		}
		err := router.doSetOne(user.OrganizationID, e.Name, e.Value)
		if err != nil {
			log.Println(err)
			if errors.Is(err, ErrAlreadyExists) {
				SendAlreadyExists(w)
			} else {
				SendInternalServerError(w)
			}
			return
		}
	}
	SendUpdated(w)
}

func (router *SettingsRouter) doSetOne(organizationID, name, value string) error {
	err := GetSettingsRepository().Set(organizationID, name, value)
	return err
}

func (router *SettingsRouter) copyToRestModel(e *OrgSetting) *GetSettingsResponse {
	m := &GetSettingsResponse{}
	m.Name = e.Name
	m.Value = e.Value
	return m
}

func (router *SettingsRouter) copyOfficeSettingsToRestModel(e *OfficeSettings) *OfficeSettingsResponse {
	return &OfficeSettingsResponse{
		WorkStartTime: e.WorkStartTime,
		WorkEndTime:   e.WorkEndTime,
	}
}

func (router *SettingsRouter) isValidSettingNameReadPublic(name string) bool {
	if name == SettingMaxBookingsPerUser.Name ||
		name == SettingMaxConcurrentBookingsPerUser.Name ||
		name == SettingMaxDaysInAdvance.Name ||
		name == SettingMaxBookingDurationHours.Name ||
		name == SettingShowNames.Name ||
		name == SettingAllowBookingsNonExistingUsers.Name ||
		name == SettingDailyBasisBooking.Name ||
		name == SettingNoAdminRestrictions.Name ||
		name == SettingCustomLogoUrl.Name ||
		name == SettingDefaultTimezone.Name ||
		name == SettingDisableBuddies.Name ||
		name == SettingFeatureRecurringBookings.Name ||
		name == SettingMaxHoursBeforeDelete.Name ||
		name == SettingEnableMaxHourBeforeDelete.Name ||
		name == SettingMinBookingDurationHours.Name ||
		name == SettingFeatureNoUserLimit.Name ||
		name == SettingFeatureCustomDomains.Name ||
		name == SettingFeatureGroups.Name ||
		name == SettingFeatureAuthProviders.Name ||
		name == SettingSubjectDefault.Name ||
		name == SysSettingOrgPrimaryDomain ||
		name == SysSettingVersion ||
		name == SysSettingDisablePasswordLogin ||
		name == SettingEnforceTOTP.Name ||
		name == SettingAllowRecurringBookings.Name {
		return true
	}
	return false
}

func (router *SettingsRouter) isValidSettingNameReadAdmin(name string) bool {
	if router.isValidSettingNameReadPublic(name) ||
		name == SettingAllowAnyUser.Name ||
		name == SettingConfluenceServerSharedSecret.Name ||
		name == SettingConfluenceAnonymous.Name ||
		name == SysSettingOrgSignupDelete ||
		name == SysSettingAdminMenuItems ||
		name == SysSettingAdminWelcomeScreens ||
		name == SettingBookingRetentionEnabled.Name ||
		name == SettingSubjectDefault.Name ||
		name == SettingBookingRetentionDays.Name ||
		name == SettingEnforceTOTP.Name ||
		name == SettingNewUserDefaultMailNotification.Name ||
		name == SettingTargetUtilizationHoursPerWeek.Name {
		return true
	}
	return false
}

func (router *SettingsRouter) isValidSettingNameWrite(name string) bool {
	if name == SettingAllowAnyUser.Name ||
		name == SettingConfluenceServerSharedSecret.Name ||
		name == SettingConfluenceAnonymous.Name ||
		name == SettingEnableMaxHourBeforeDelete.Name ||
		name == SettingMaxBookingsPerUser.Name ||
		name == SettingMaxConcurrentBookingsPerUser.Name ||
		name == SettingMaxDaysInAdvance.Name ||
		name == SettingMaxHoursBeforeDelete.Name ||
		name == SettingMinBookingDurationHours.Name ||
		name == SettingDailyBasisBooking.Name ||
		name == SettingNoAdminRestrictions.Name ||
		name == SettingCustomLogoUrl.Name ||
		name == SettingShowNames.Name ||
		name == SettingAllowBookingsNonExistingUsers.Name ||
		name == SettingMaxBookingDurationHours.Name ||
		name == SettingDisableBuddies.Name ||
		name == SettingDefaultTimezone.Name ||
		name == SettingBookingRetentionDays.Name ||
		name == SettingBookingRetentionEnabled.Name ||
		name == SettingAllowRecurringBookings.Name ||
		name == SettingNewUserDefaultMailNotification.Name ||
		name == SettingEnforceTOTP.Name ||
		name == SettingSubjectDefault.Name ||
		name == SettingTargetUtilizationHoursPerWeek.Name {
		return true
	}
	return false
}

func (router *SettingsRouter) getSettingType(name string) SettingType {
	if name == SettingAllowAnyUser.Name {
		return SettingAllowAnyUser.Type
	}
	if name == SettingConfluenceServerSharedSecret.Name {
		return SettingConfluenceServerSharedSecret.Type
	}
	if name == SettingConfluenceAnonymous.Name {
		return SettingConfluenceAnonymous.Type
	}
	if name == SettingMaxBookingsPerUser.Name {
		return SettingMaxBookingsPerUser.Type
	}
	if name == SettingMaxConcurrentBookingsPerUser.Name {
		return SettingMaxConcurrentBookingsPerUser.Type
	}
	if name == SettingMaxDaysInAdvance.Name {
		return SettingMaxDaysInAdvance.Type
	}
	if name == SettingMaxBookingDurationHours.Name {
		return SettingMaxBookingDurationHours.Type
	}
	if name == SettingDailyBasisBooking.Name {
		return SettingDailyBasisBooking.Type
	}
	if name == SettingNoAdminRestrictions.Name {
		return SettingNoAdminRestrictions.Type
	}
	if name == SettingShowNames.Name {
		return SettingShowNames.Type
	}
	if name == SettingAllowBookingsNonExistingUsers.Name {
		return SettingAllowBookingsNonExistingUsers.Type
	}
	if name == SettingDefaultTimezone.Name {
		return SettingDefaultTimezone.Type
	}
	if name == SettingMaxHoursBeforeDelete.Name {
		return SettingMaxHoursBeforeDelete.Type
	}
	if name == SettingEnableMaxHourBeforeDelete.Name {
		return SettingEnableMaxHourBeforeDelete.Type
	}
	if name == SettingDisableBuddies.Name {
		return SettingDisableBuddies.Type
	}
	if name == SettingCustomLogoUrl.Name {
		return SettingCustomLogoUrl.Type
	}
	if name == SettingMinBookingDurationHours.Name {
		return SettingMinBookingDurationHours.Type
	}
	if name == SettingTargetUtilizationHoursPerWeek.Name {
		return SettingTargetUtilizationHoursPerWeek.Type
	}
	if name == SettingBookingRetentionDays.Name {
		return SettingBookingRetentionDays.Type
	}
	if name == SettingBookingRetentionEnabled.Name {
		return SettingBookingRetentionEnabled.Type
	}
	if name == SettingAllowRecurringBookings.Name {
		return SettingAllowRecurringBookings.Type
	}
	if name == SettingNewUserDefaultMailNotification.Name {
		return SettingNewUserDefaultMailNotification.Type
	}
	if name == SettingEnforceTOTP.Name {
		return SettingEnforceTOTP.Type
	}
	if name == SettingSubjectDefault.Name {
		return SettingSubjectDefault.Type
	}
	return 0
}

func (router *SettingsRouter) isValidSettingType(name string, value string) bool {
	settingType := router.getSettingType(name)
	if settingType == 0 {
		return false
	}
	if settingType == SettingTypeString && len(value) <= 256 {
		return true
	}
	if settingType == SettingTypeBool && (value == "1" || value == "0") {
		return true
	}
	if settingType == SettingTypeInt {
		if _, err := strconv.Atoi(value); err == nil {
			return true
		}
	}
	return false
}

func (router *SettingsRouter) isValidSettingValue(name string, value string) bool {
	if value == "" {
		return true
	}
	if name == SettingCustomLogoUrl.Name {
		if !ValidateURL(value) {
			return false
		}
		return true
	}
	if name == SettingMaxBookingsPerUser.Name {
		if !ValidateNumber(value, 1, 9999) {
			return false
		}
		return true
	}
	if name == SettingMaxConcurrentBookingsPerUser.Name || name == SettingMaxDaysInAdvance.Name || name == SettingMaxHoursBeforeDelete.Name ||
		name == SettingMaxBookingDurationHours.Name || name == SettingMinBookingDurationHours.Name {
		if !ValidateNumber(value, 0, 9999) {
			return false
		}
		return true
	}
	if name == SettingBookingRetentionDays.Name {
		if !ValidateNumber(value, 30, 999) {
			return false
		}
		return true
	}
	if name == SettingTargetUtilizationHoursPerWeek.Name {
		if !ValidateNumber(value, 0, 168) {
			return false
		}
		return true
	}
	if name == SettingDefaultTimezone.Name && !IsValidTimeZone(value) {
		return false
	}
	if name == SettingConfluenceServerSharedSecret.Name && len(value) > 256 {
		return false
	}
	if name == SettingSubjectDefault.Name {
		intVal, _ := strconv.Atoi(value)
		if intVal != SettingSubjectDefaultDisabled &&
			intVal != SettingSubjectDefaultOptional &&
			intVal != SettingSubjectDefaultRequired {
			return false
		}
	}
	return true
}

func (router *SettingsRouter) getAdminWelcomeScreens(settings []*OrgSetting) *GetSettingsResponse {
	res := []SettingsRouterWelcomeScreen{}
	for _, plg := range plugin.GetPlugins() {
		ws := (*plg).GetAdminWelcomeScreen()
		if ws != nil {
			skip := false
			for _, setting := range settings {
				if setting.Name == ws.SkipOnSettingTrue && setting.Value == "1" {
					skip = true
					break
				}
			}
			if !skip {
				resItem := SettingsRouterWelcomeScreen{
					Source: ws.Source,
				}
				res = append(res, resItem)
			}
		}
	}
	jsonBytes, err := json.Marshal(res)
	if err != nil {
		log.Println("Error marshalling welcome screens:", err)
		return nil
	}
	return &GetSettingsResponse{
		Name:  SysSettingAdminWelcomeScreens,
		Value: string(jsonBytes),
	}
}

func (router *SettingsRouter) getAdminMenuItems() *GetSettingsResponse {
	res := []SettingsRouterAdminMenuItem{}
	for _, plg := range plugin.GetPlugins() {
		for _, item := range (*plg).GetAdminUIMenuItems() {
			resItem := SettingsRouterAdminMenuItem{
				ID:         item.ID,
				Title:      item.Title,
				Source:     item.Source,
				Visibility: item.Visibility,
				Icon:       item.Icon,
			}
			res = append(res, resItem)
		}
	}
	jsonBytes, err := json.Marshal(res)
	if err != nil {
		log.Println("Error marshalling admin menu items:", err)
		return nil
	}
	return &GetSettingsResponse{
		Name:  SysSettingAdminMenuItems,
		Value: string(jsonBytes),
	}
}

func (router *SettingsRouter) getSysSettingOrgSignupDelete() *GetSettingsResponse {
	boolVal := "0"
	if GetConfig().AllowOrgDelete {
		boolVal = "1"
	}
	return &GetSettingsResponse{
		Name:  SysSettingOrgSignupDelete,
		Value: boolVal,
	}
}

func (router *SettingsRouter) getSysSettingVersion() *GetSettingsResponse {
	return &GetSettingsResponse{
		Name:  SysSettingVersion,
		Value: GetProductVersion(),
	}
}

func (router *SettingsRouter) getSysSettingDisablePasswordLogin() *GetSettingsResponse {
	boolVal := "0"
	if GetConfig().DisablePasswordLogin {
		boolVal = "1"
	}
	return &GetSettingsResponse{
		Name:  SysSettingDisablePasswordLogin,
		Value: boolVal,
	}
}

func (router *SettingsRouter) getSysSettingOrgPrimaryDomain(orgId string) *GetSettingsResponse {
	org, _ := GetOrganizationRepository().GetOne(orgId)
	primaryDomain, _ := GetOrganizationRepository().GetPrimaryDomain(org)

	primaryDomainValue := ""
	if primaryDomain != nil {
		primaryDomainValue = primaryDomain.DomainName
	}

	return &GetSettingsResponse{
		Name:  SysSettingOrgPrimaryDomain,
		Value: primaryDomainValue,
	}
}