# specs/openapi.yaml
openapi: 3.1.0
info:
  title: Seatsurfing REST API
  description: |
    Seatsurfing provides REST APIs used by the mobile and web app as well as the administrator web interface.

    ## Authorization checks

    Each API endpoint is protected by authorization checks:
    1. **Role check**: Role of the user (Anonymous, User, Space Admin, Org Admin, Super Admin)
    2. **Organization check**: Organization membership

    ## Roles

    - **Anonymous**: Every user allowed, no authorization required
    - **User**: Authorized user who is a member of an organization
    - **Space Admin**: An organization's space administrator (can manage locations, spaces, bookings)
    - **Org Admin**: An organization's full administrator
    - **Super Admin**: Global administrator with access to all organizations and all functions

    ## Authentication

    Authentication is done via JWT Bearer tokens in the `Authorization` header. Service accounts may also use HTTP Basic Authentication.

    ## Session management

    - Access Token Lifetime: 15 minutes
    - Refresh Token Lifetime: 28 days
    - Tokens are signed using RS512

    ## Custom error codes

    Some endpoints return custom error codes via the `X-Error-Code` response header:
    - `1001`: Booking slot conflict
    - `1002`: Location max concurrent bookings reached
    - `1003`: Too many upcoming bookings for user
    - `1004`: Too many days in advance
    - `1005`: Invalid booking duration
...
    - `1010`: Booking subject required
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained via login or refresh endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: Service account authentication using `{orgId}_{email}` as username and the service account password
  schemas:
...
    OfficeSettings:
      type: object
      description: Singleton `office_settings` model used to configure daily office hours.
      required: [workStartTime, workEndTime]
      properties:
        id:
          type: integer
          enum: [1]
          description: Singleton row ID
        workStartTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "08:00"
          description: Maps to `office_settings.work_start_time`; must be before `workEndTime`
        workEndTime:
          type: string
          pattern: "^([01]\\d|2[0-3]):[0-5]\\d$"
          example: "17:00"
          description: Maps to `office_settings.work_end_time`
...
    SetSettingsRequest:
      type: object
      properties:
        value:
          type: string
    GetSettingsResponse:
      type: object
      properties:
        name:
          type: string
        value:
          type: string

    # --- Statistics ---
...
  responses:
...
    Updated:
      description: Object updated or deleted successfully (204 No Content)
    BadRequest:
      description: Malformed request or incomplete JSON
      headers:
        X-Error-Code:
          schema:
            type: integer
          description: Custom error code (see error codes table)
...
    Forbidden:
      description: Authorized, but user has no access to requested resource
    NotFound:
      description: Object not found
    Conflict:
      description: Conflicting request (e.g. slot already booked)
...
paths:
...
  /setting/:
    get:
      tags: [Settings]
      summary: Get all settings
      description: |
        Returns all organization settings visible to the current user.
        Regular users see public settings; Org Admins see additional admin settings;
        Space Admins see admin menu items. Also includes system settings and plugin settings.
      operationId: getAllSettings
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of settings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GetSettingsResponse"
        "403":
          $ref: "#/components/responses/Forbidden"
    put:
      tags: [Settings]
      summary: Update all settings
      description: Updates multiple settings at once. Requires Org Admin role.
      operationId: setAllSettings
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/GetSettingsResponse"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Invalid setting name
        "409":
          $ref: "#/components/responses/Conflict"
  /setting/office-hours:
    get:
      tags: [Settings]
      summary: Get office hours
      description: Returns the global office-hours configuration. Requires Org Admin role.
      operationId: getOfficeHours
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Office-hours configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfficeSettings"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Settings]
      summary: Update office hours
      description: Creates or updates the global office-hours configuration. Requires Org Admin role.
      operationId: setOfficeHours
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfficeSettings"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /setting/{name}:
    get:
      tags: [Settings]
      summary: Get a setting
      description: Returns a single setting by name. Access depends on the setting type (public vs admin).
      operationId: getSetting
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Setting name
      responses:
        "200":
          description: Setting value
          content:
            application/json:
              schema:
                type: string
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Settings]
      summary: Update a setting
      description: Updates a single setting by name. Requires Org Admin role.
      operationId: setSetting
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
          description: Setting name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetSettingsRequest"
      responses:
        "204":
          $ref: "#/components/responses/Updated"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Invalid setting name
        "409":
          $ref: "#/components/responses/Conflict"
  /setting/timezones:
    get:
      tags: [Settings]
      summary: Get available timezones
      description: Returns the list of all available IANA timezones.
      operationId: getTimezones
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of timezone identifiers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string

  # ===========================
  # User Preferences
  # ===========================
...